change fragment shader color varying to lowp

+++ packages/wave-gradients/src/shaders/color.frag
@@ -1,4 +1,4 @@
-varying highp vec3 color;
+varying lowp vec3 color;
 
 void main() {
   gl_FragColor = vec4(color, 1.0);
